Remove support for old version (per-8.4) of postgresql
The _fix_timezone sub had a workaround for a bug in postgresql's
handling of pre-1901 an post-2038 dates. This is no longer needed.
The bug was fixed in pg 8.4. The fix is to remove the work around
and the corresponding test. As we get closer to 2038 this will
become more apparent.